requirements.txt
PyYAML
jsonschema
watchdog
toml
lxml